# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Todo
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /todos/{id}:
    x-swagger-router-controller: todos
    get:
      description: get a todo item
      operationId: show
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTodoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a todo
      operationId: update
      parameters:
        - name: id
          description: todo item id
          type: string
          in: path
          required: true
        - name: title
          description: todo properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Todo"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a todo
      operationId: destroy
      parameters:
        - name: id
          description: todo item id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Todo:
    type: object
    properties:
      title:
        type: string
        description: Title of the todo item
    required:
        - title
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  GetTodoResponse:
    required:
      - id
      - title
    properties:
      id:
        type: string
      title: 
        type: string